# ADBDEV-7158: Do not use CdbDispatchCommand() in pg_lock_status()