--- a/.abi-check/6.27.1_arenadata61/postgres.symbols.ignore
+++ b/.abi-check/6.27.1_arenadata61/postgres.symbols.ignore
@@ -1,0 +1,1 @@
+ConfigureNamesBool_gp