Do not use CdbDispatchCommand() in pg_lock_status()

Previosly, `pg_lock_status()` implementation used `CdbDispatchCommand()` which
relied on libpq's `pqRowProcessor()` that internally performed humongous amount
of `malloc()`/`realloc()` allocations for large queries. This caused progressive
memory leaks that could exhaust system memory if querying `pg_locks` frequently
or with many concurrent locks.

This patch completely replaces `CdbDispatchCommand()` in `pg_lock_status()` with
executor-based routines that stream row-by-row directly through the executor
infrastructure, while also improving its performance.

Benchmarks show the changes reduce peak memory usage by 63% (down to 350MB) and
improve execution time by 20% (from 7.6s to 6.1s) for the same 1.5 million lock
test case.

The results are still materialized, unfortunately. This occurs because
`pg_locks` view selects everything from `pg_lock_status()`, regardless of what
LIMIT clause has. The way to bypass it is to select `pg_lock_status()` directly:

    SELECT pg_lock_status() LIMIT 10;

---
Commit 7f6d6c6 increased the size of the `ConfigureNamesBool_gp` structure and added it to the ignore list for the `6.27.1_arenadata60` tag, but now the current tag is `6.27.1_arenadata61`, and that commit is not in it yet, so I had to add this structure to the ignore list for the current tag as well.

---
### Profiling

Preparation, in a separate session:
```sql
CREATE TABLE t_part_small(i int, j int) PARTITION BY RANGE (j) (START (0) END (50) EVERY (1)); 
BEGIN;
LOCK TABLE t_part_small IN ACCESS EXCLUSIVE MODE;
```

#### Without the patch
`CdbDispatchCommand(...)`'s query is replaced with: 
```sql
select l.* from pg_catalog.pg_locks l, generate_series(1, 10000)
```

```
postgres=# select count(*) from pg_locks;
  count  
---------
 1590055
(1 row)
postgres=# 

Time: 7601.848 ms
postgres=# select pg_backend_pid();
 pg_backend_pid 
----------------
          10323
(1 row)

postgres=# \!grep VmPeak /proc/10323/status
VmPeak:	 975360 kB -- (~950 MB)
```

#### With patch
`PG_LOCKS_INTERNAL_QUERY` macro is replaced with: 
```sql
select l.* from gp_dist_random('pg_catalog.pg_locks') l, generate_series(1, 10000)
```

```
postgres=# select count(*) from pg_locks;
  count  
---------
 1590055
(1 row)

Time: 6137.254 ms
postgres=# select pg_backend_pid();
 pg_backend_pid 
----------------
          20436
(1 row)

postgres=# \!grep VmPeak /proc/20436/status
VmPeak:	 357580 kB -- (~350 MB)
```