@RekGRpth 2025-03-31
> select l.* from gp_dist_random('pg_catalog.pg_locks') l, generate_series(1, 10000)

Is it possible to make a test out of this using injection fault?